+++ integration-tests/src/test.rs
@@ -2,10 +2,18 @@
 mod tests {
     extern crate server_lib;
     extern crate client_lib;
+    extern crate bitcoin;
 
+    use client_lib::wallet::Wallet;
     use client_lib::*;
     use server_lib::server;
+
+    use bitcoin::{ Amount, TxIn };
+    use bitcoin::OutPoint;
+    use bitcoin::hashes::sha256d;
+
     use std::{thread, time};
+    use std::collections::HashMap;
 
     #[test]
     fn test_ecdsa() {
@@ -61,7 +69,37 @@ mod tests {
         spawn_server();
 
         let client_shim = ClientShim::new("http://localhost:8000".to_string(), None);
-        let resp = state_entity::deposit::deposit(&client_shim);
+
+        // 2P-ECDSA
+        let private_share: ecdsa::PrivateShare = ecdsa::get_master_key(&client_shim);
+        println!("{}",private_share.id);
+        println!("{:?}",private_share.master_key.public);
+
+        // make wallet
+        let id = String::from("1");
+        let last_derived_pos = 0;
+        let addresses_derivation_map = HashMap::new();
+        let network = "regtest".to_string();
+        let mut wallet = Wallet {
+            id,
+            network,
+            private_share,
+            last_derived_pos,
+            addresses_derivation_map,
+        };
+
+
+        let addr = wallet.get_new_bitcoin_address();
+        let amount = Amount::ONE_BTC;
+        let inputs =  vec![
+            TxIn {
+                previous_output: OutPoint { txid: sha256d::Hash::default(), vout: 0 },
+                sequence: 0xffffffff - 2,
+                witness: Vec::new(),
+                script_sig: bitcoin::Script::default(),
+            }
+        ];
+        let resp = state_entity::deposit::deposit(wallet, &client_shim, inputs, addr, amount);
         println!("{} {}",resp.0,resp.1);
         assert!(resp.0 == String::from("deposit"));
     }
